added dragging and dropping items in inventory and stack splitting using the scrollwheel, made it so you can only sprint if moving forward

+++ Assets/Item.cs
@@ -7,7 +7,7 @@ using UnityEngine.UI;
 
 public class Item : MonoBehaviour
 {
-    public float maxStackCount;
+    public int maxStackCount;
     public bool isStackable = false;
     public GameObject icon;
 }
